% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_songs.R
\name{create_inputfiles}
\alias{create_inputfiles}
\title{Create the rmd files}
\usage{
create_inputfiles(
  input = ".",
  songbook_yamlfile = file.path(input, "_songbookdown.yml"),
  bookdown_yamlfile = file.path(input, "_bookdown.yml")
)
}
\arguments{
\item{input}{a folder}

\item{songbook_yamlfile}{the songbook yaml file containing the songbook metadata}

\item{bookdown_yamlfile}{the bookdown yaml file containing the bookdown metadata}
}
\value{
A batch of Rmd Files which can be used to create the songbook.
}
\description{
Create the rmd files
}
