% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template}
\alias{create_template}
\title{Create a songbook template}
\usage{
create_template(folder = ".")
}
\arguments{
\item{folder}{optional: path to folder where this template should be created. Defaults to the cwd}
}
\value{
does not return anything, simply creates files in the specified directory
}
\description{
Create a songbook template
}
