% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_songs.R
\name{read_yamlheader}
\alias{read_yamlheader}
\title{Read yaml header}
\usage{
read_yamlheader(input, isfile = TRUE)
}
\arguments{
\item{input}{Path to a text file (if \code{isfile} is \code{TRUE}) or a character vector (output from \code{readLines()})}

\item{isfile}{boolean, indicating whether \code{input} is a file or a character vector}
}
\value{
A named list
}
\description{
Reads a yaml header from a text file or character vector (latter is the output of \code{readLines()})
}
