% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_edit_url.R
\name{replace_edit_url}
\alias{replace_edit_url}
\title{Replace the (github) "Edit"-URLs}
\usage{
replace_edit_url(
  songbook_yamlfile = list.files(pattern = "_songbookdown.ya*ml"),
  bookdown_yamlfile = list.files(pattern = "_bookdown.ya*ml")
)
}
\arguments{
\item{songbook_yamlfile, bookdown_yamlfile}{the songbook- / bookdown- yamlfile containing
yaml the songbook metadata. Must be relative to the input.}
}
\value{
Does not return a value, simply modiefies the html files.
}
\description{
This function replaces the URL that normally links to the Rmd File on github
with the URL of the source text file.
}
