% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_songs.R
\name{yaml_headerpos}
\alias{yaml_headerpos}
\title{Get yaml header poition}
\usage{
yaml_headerpos(input)
}
\arguments{
\item{input}{a character vector}
}
\value{
an integer vector of legnth 2
}
\description{
Gets the position of the yaml inline header by searching for the first
two occurences of "---" (three or more dashes accepted)
}
