% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_songs.R
\name{create_inputfiles}
\alias{create_inputfiles}
\title{Create the rmd files}
\usage{
create_inputfiles(
  input = getwd(),
  songbook_yamlfile = list.files(pattern = "_songbookdown.ya*ml"),
  bookdown_yamlfile = list.files(pattern = "_bookdown.ya*ml")
)
}
\arguments{
\item{input}{a folder}

\item{songbook_yamlfile, bookdown_yamlfile}{the songbook- / bookdown- yamlfile containing
yaml the songbook metadata. Must be relative to the input.}
}
\value{
A batch of Rmd Files which can be used to create the songbook.
}
\description{
Create the rmd files
}
